requirements.txt
flask
psycopg2-binary
pandas
scikit-learn
